Move classes from root\entity\dto to root\dto
