+++ src/main/java/com/fmachinus/practice/dto/CustomerDto.java
@@ -1,7 +1,6 @@
-package com.fmachinus.practice.entity.dto;
+package com.fmachinus.practice.dto;
 
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 
 import java.math.BigDecimal;
