+++ src/main/java/com/fmachinus/practice/dto/OrderDto.java
@@ -1,4 +1,4 @@
-package com.fmachinus.practice.entity.dto;
+package com.fmachinus.practice.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
